Move code to discover modules
