+++ src/tiledb/cloud/bioimg/__init__.py
@@ -1,4 +1,16 @@
-from tiledb.cloud.bioimg.exportation import export
-from tiledb.cloud.bioimg.ingestion import ingest
+from .exportation import export
+from .ingestion import ingest
+from .helpers import get_embeddings_uris
+from .helpers import serialize_filter
+from .helpers import batch
+from .helpers import scale_calc
 
-__all__ = ("ingest", "export")
+# from .types import SupportedExtensions, EMBEDDINGS
+
+__all__ = ("ingest", 
+           "export",
+           "get_embeddings_uris",
+           "serialize_filter",
+           "batch"
+           "scale_calc"
+           )
